feat(gimbal): Gimbal integration with interlock, ballistics, and emergency stop
Completes Subsystem 3 of the system development plan with:

- Wired AuroreLinkServer callbacks for mode switching (AUTO/FREECAM) and operator
  authorization, enabling remote gimbal control via the frontend
- Rewrote actuation thread with:
  * Interlock safety gating: only send servo commands when interlock allows AND
    system is in TRACKING/ARMED/FREECAM states
  * GimbalController integration: converts track centroid to servo angles in AUTO
    mode, accepts absolute angles in FREECAM mode
  * BallisticSolver integration: computes fire control solutions with p_hit lookup
    tables; feeds solutions to state machine for ARMED mode gating
  * Actual gimbal feedback: replaces pixel-proxy with real servo angle readback
    from FusionHat.get_servo_angle()
  * Throttled telemetry broadcast (30Hz) with gimbal status, ballistics data, and
    FCS state for HUD display
- Added emergency stop in shutdown: disable_all_servos() + set_inhibit(true) +
  safety telemetry event
- Added interlock watchdog feed in safety monitor thread (1kHz)
- Moved telemetry broadcasting from track thread to actuation thread for better
  synchronization with gimbal commands

Verified: dry-run mode starts cleanly, logs SYSTEM_BOOT event, runs 4-thread
pipeline, and shuts down with emergency stop. All 18 active unit tests pass
(1 pre-existing test failure in SafetyMonitorTest unrelated to this change).

Fixes AM7-L2-VIS-008 (gimbal control), AM7-L3-MODE-007 (safety interlock),
AM7-L3-FIRE-001 (ballistics), and AM7-L2-TIM-001 (telemetry timing).

+++ include/aurore/state_machine.hpp
@@ -73,8 +73,8 @@ enum class FcsState : uint8_t {
 struct Detection {
     float confidence{0.0f};
     struct { int x, y, w, h; } bbox{};
-    float cx() const { return static_cast<float>(bbox.x) + bbox.w * 0.5f; }
-    float cy() const { return static_cast<float>(bbox.y) + bbox.h * 0.5f; }
+    float cx() const { return static_cast<float>(bbox.x) + static_cast<float>(bbox.w) * 0.5f; }
+    float cy() const { return static_cast<float>(bbox.y) + static_cast<float>(bbox.h) * 0.5f; }
 };
 
 // Note: Named GimbalStatusSm to avoid conflict with proto-generated aurore::GimbalStatus
